move config sections to own package
