+++ src/client/java/com/github/thedeathlycow/scorchful/client/ShaderStatusEffectManagers.java
@@ -1,7 +1,7 @@
 package com.github.thedeathlycow.scorchful.client;
 
 import com.github.thedeathlycow.scorchful.Scorchful;
-import com.github.thedeathlycow.scorchful.config.ClientConfig;
+import com.github.thedeathlycow.scorchful.config.section.ClientConfig;
 import com.github.thedeathlycow.scorchful.registry.SStatusEffects;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.minecraft.client.network.ClientPlayerEntity;
