+++ src/main/java/com/github/thedeathlycow/scorchful/config/section/CombatConfig.java
@@ -1,6 +1,8 @@
-package com.github.thedeathlycow.scorchful.config;
+package com.github.thedeathlycow.scorchful.config.section;
 
 import com.github.thedeathlycow.scorchful.Scorchful;
+import com.github.thedeathlycow.scorchful.config.ScorchfulConfig;
+import com.github.thedeathlycow.scorchful.config.Translate;
 import com.github.thedeathlycow.scorchful.item.FireChargeThrower;
 import dev.isxander.yacl3.config.v2.api.ConfigClassHandler;
 import dev.isxander.yacl3.config.v2.api.SerialEntry;
