+++ src/main/java/com/github/thedeathlycow/scorchful/config/section/HeatingConfig.java
@@ -1,6 +1,8 @@
-package com.github.thedeathlycow.scorchful.config;
+package com.github.thedeathlycow.scorchful.config.section;
 
 import com.github.thedeathlycow.scorchful.Scorchful;
+import com.github.thedeathlycow.scorchful.config.ScorchfulConfig;
+import com.github.thedeathlycow.scorchful.config.Translate;
 import dev.isxander.yacl3.config.v2.api.ConfigClassHandler;
 import dev.isxander.yacl3.config.v2.api.SerialEntry;
 import dev.isxander.yacl3.config.v2.api.autogen.*;
