+++ src/main/java/com/github/thedeathlycow/scorchful/item/TurtleArmorEffects.java
@@ -1,7 +1,7 @@
 package com.github.thedeathlycow.scorchful.item;
 
 import com.github.thedeathlycow.scorchful.Scorchful;
-import com.github.thedeathlycow.scorchful.config.HeatingConfig;
+import com.github.thedeathlycow.scorchful.config.section.HeatingConfig;
 import com.github.thedeathlycow.scorchful.config.ScorchfulConfig;
 import com.github.thedeathlycow.scorchful.registry.SEntityAttributes;
 import com.github.thedeathlycow.scorchful.registry.SItems;
