+++ src/main/java/com/github/thedeathlycow/scorchful/item/component/DrinkLevelComponent.java
@@ -1,11 +1,10 @@
 package com.github.thedeathlycow.scorchful.item.component;
 
-import com.github.thedeathlycow.scorchful.Scorchful;
 import com.github.thedeathlycow.scorchful.api.ServerThirstPlugin;
 import com.github.thedeathlycow.scorchful.components.PlayerWaterComponent;
 import com.github.thedeathlycow.scorchful.components.ScorchfulComponents;
 import com.github.thedeathlycow.scorchful.config.ScorchfulConfig;
-import com.github.thedeathlycow.scorchful.config.ThirstConfig;
+import com.github.thedeathlycow.scorchful.config.section.ThirstConfig;
 import com.github.thedeathlycow.scorchful.item.WaterSkinItem;
 import com.github.thedeathlycow.scorchful.mixin.accessor.RegistryEntryReferenceAccessor;
 import com.github.thedeathlycow.scorchful.registry.SDataComponentTypes;
