+++ src/main/java/com/github/thedeathlycow/scorchful/server/SandAccumulation.java
@@ -1,9 +1,8 @@
 package com.github.thedeathlycow.scorchful.server;
 
-import com.github.thedeathlycow.scorchful.Scorchful;
 import com.github.thedeathlycow.scorchful.block.SandPileBlock;
 import com.github.thedeathlycow.scorchful.config.ScorchfulConfig;
-import com.github.thedeathlycow.scorchful.config.WeatherConfig;
+import com.github.thedeathlycow.scorchful.config.section.WeatherConfig;
 import com.github.thedeathlycow.scorchful.registry.SBlocks;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
